pyproject.toml
[project]
name = "dogverificationsystem"
version = "0.1.0"
description = "Dog Verification System for Kaggle P100"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "datasets>=2.1.0",
    "einops>=0.8.0",
    "hydra-core>=1.3.2",
    "librosa>=0.10.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "soundfile>=0.12.1",
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "torchvision==0.19.1",
    "torchmetrics>=1.4.0",
    "tqdm>=4.66.0",
    "wandb>=0.16.0",
    "yadisk>=3.4.0",
    "pytubefix>=10.3.8",
    "speechbrain>=1.1.0",
    "torchcodec>=0.11.0",
    "kagglehub>=1.0.0",
    "opencv-python>=4.13.0.92",
    "pydub>=0.25.1",
    "seaborn>=0.13.2",
]
